refactor(ctrl): migrate GitHub webhook onto pkg/webhook receiver
